NVIDIA/benchmarks/flux1/implementations/theia_ngpu16_ngc25.09_nemo/requirements.txt
git+https://github.com/mlcommons/logging.git@5.1.0-rc3
git+https://github.com/NVIDIA/mlperf-common.git@b86d175a05849d650a8ff69c1e2c37b9f4e61d51
megatron-energon==7.1.0
# WAR for MultiStorageClient
multi-storage-client<0.26.0
